requirements.txt
gradio
google-generativeai
python-dotenv
tldextract
html2text
playwright
scikit-learn
